Change default window size

+++ ImageUtils.py
@@ -11,7 +11,7 @@ def rgb2grayscale(img: QImage):
     if bytesPerPixel == 1:
         return img
 
-    arr = np.asarray(ptr).reshape((img.height(), img.width(), bytesPerPixel))
+    arr = np.asarray(ptr, dtype=np.uint8).reshape((img.height(), img.width(), bytesPerPixel))
     mat = [0.2989, 0.5870, 0.1140, 0] if bytesPerPixel == 4 else [0.2989, 0.5870, 0.1140]
     grayArr = np.array(np.dot(arr, mat), dtype=np.uint8)
